feat: split messages.go into commands.go and queries.go for improved organization
